refactor(web): serve internal job lookup via core
The internal `/api/internal/jobs/[jobId]` route was the last web API
route reading the database directly (`jobRepository.getJobById` +
Prisma). It now proxies the session-authenticated core endpoint
`GET /v1/jobs/{id}` instead, matching the SSR job-details loader.

To keep the route's response contract (the job share modal reads
`job.share` from the cached job across refetches), core's job details
schema now includes the already-loaded `share` relation:

- move `jobShareSchema`/`taskShareSchema` to `schemas/share.schema.ts`
  so `job.schema.ts` can reference them without a circular import
- add `share` to `jobSchema` (as a union with null so the generated
  `JobShare` component stays non-nullable) and `serializeJobDetails`
- map `job.share` in `mapCoreJobToJobWithSokosumiStatus`, relocating
  `mapCoreJobShare` into `core-dto-mappers` to avoid an import cycle
- regenerate the web core client snapshot

Authorization now follows core's job read rules (workspace-scoped via
`requireJobReadForRouteVars`) instead of strict job ownership, aligning
client refetches with the server-rendered job details page, which
already renders read-only for non-owner workspace members.
