+++ apps/core/src/routes/v1/agents/[id]/demo-jobs/post.test.ts
@@ -87,6 +87,7 @@ function createSerializedJob(overrides: Record<string, unknown> = {}) {
         links: [],
       },
     ],
+    share: null,
     ...overrides,
   };
 }
