+++ apps/core/src/routes/v1/jobs/[id]/workspace/put.test.ts
@@ -139,6 +139,7 @@ function createJobApi(overrides: Partial<Record<string, unknown>> = {}) {
       overrideLegalTerms: null,
     },
     events: [],
+    share: null,
     ...overrides,
   };
 }
